Refactored the way to get layer's metadata, decoupled and easier to maintain
