+++ lib/windshaft/backends/map.js
@@ -1,16 +1,11 @@
-var assert = require('assert');
+'use strict';
 
-var _ = require('underscore');
-var queue = require('queue-async');
+var assert = require('assert');
 var step = require('step');
 var debug = require('debug')('windshaft:backend:map');
-var PSQL = require('cartodb-psql');
-var RendererParams = require('../renderers/renderer_params');
-var queryUtils = require('../utils/query_utils');
-
-var DummyMapConfigProvider = require('../models/providers/dummy_mapconfig_provider');
 var Timer = require('../stats/timer');
 
+var layerMetadata = require('../metadata')();
 
 /**
  * @param {RendererCache} rendererCache
@@ -89,32 +84,33 @@ MapBackend.prototype.createLayergroup = function(mapConfig, params, validatorMap
 
             var next = this;
 
-            getLayersMetadata(self._rendererCache, params, mapConfig, function(err, layersMetadata) {
+            layerMetadata.getMetadata(self._rendererCache, params, mapConfig, function(err, layersMetadata) {
                 if (err) {
-                    self._mapStore.del(mapConfig.id(), function(delErr) {
+                    return self._mapStore.del(mapConfig.id(), function(delErr) {
                         if (delErr) {
                             debug("Failed to delete MapConfig '" + mapConfig.id() + " after: " + err);
                         }
                         return next(err);
                     });
-                } else {
-                    if (layersMetadata) {
-                        response.metadata = response.metadata || {};
-                        response.metadata.layers = layersMetadata;
-
-                        // backwards compatibility for torque
-                        var torqueMetadata = layersMetadata.reduce(function(acc, layer, layerId) {
-                            if (layer.type === 'torque') {
-                                acc[layerId] = layer.meta;
-                            }
-                            return acc;
-                        }, {});
-                        if (Object.keys(torqueMetadata).length) {
-                            response.metadata.torque = torqueMetadata;
-                        }
-                    }
-                    return next();
                 }
+
+                if (layersMetadata) {
+                    response.metadata = response.metadata || {};
+                    response.metadata.layers = layersMetadata;
+
+                    // backwards compatibility for torque
+                    var torqueMetadata = layersMetadata.reduce(function(acc, layer, layerId) {
+                        if (layer.type === 'torque') {
+                            acc[layerId] = layer.meta;
+                        }
+                        return acc;
+                    }, {});
+                    if (Object.keys(torqueMetadata).length) {
+                        response.metadata.torque = torqueMetadata;
+                    }
+                }
+
+                return next();
             });
         },
         function finish(err) {
@@ -123,156 +119,3 @@ MapBackend.prototype.createLayergroup = function(mapConfig, params, validatorMap
         }
     );
 };
-
-function getLayersMetadata(rendererCache, params, mapConfig, callback) {
-    var metadata = [];
-
-    var metaQueue = queue(mapConfig.getLayers().length);
-
-    mapConfig.getLayers().forEach(function(layer, layerId) {
-        var layerType = mapConfig.layerType(layerId);
-
-        if (layerType === 'torque') {
-            metaQueue.defer(getTorqueLayerMetadata, rendererCache, mapConfig, params, 'json.torque', layerId);
-        } else if (layerType === 'mapnik') {
-            metaQueue.defer(getMapnikLayerMetadata, layer, mapConfig, params);
-        } else {
-            metaQueue.defer(getEmptyLayerMetadada);
-        }
-    });
-
-    metaQueue.awaitAll(function (err, results) {
-        if (err) {
-            return callback(err);
-        }
-
-        if (!results) {
-            return callback(null, null);
-        }
-
-        mapConfig.getLayers().forEach(function(layer, layerId) {
-            var layerType = mapConfig.layerType(layerId);
-
-            metadata[layerId] = {
-                type: layerType,
-                meta: results[layerId]
-            };
-
-            var widgets = mapConfig.getWidgets(layerId);
-
-            if (_.size(widgets)) {
-                metadata[layerId].widgets = widgets;
-            }
-        });
-
-        return callback(err, metadata);
-    });
-}
-
-/// Fetch metadata for a tileset in a MapConfig
-//
-/// @param rendererType a MapConfig renderer type (layer.type in MapConfig spec)
-/// @param layerId layer index within the mapConfig
-/// @param callback function(err, metadata) where metadata format
-///
-function getTorqueLayerMetadata(rendererCache, mapConfig, params, format, layerId, callback) {
-    params = _.extend({}, params, {
-        token: mapConfig.id(),
-        format: format,
-        layer: layerId
-    });
-
-    var renderer;
-
-    step(
-        function(){
-            rendererCache.getRenderer(new DummyMapConfigProvider(mapConfig, params), this);
-        },
-        function(err, _renderer) {
-            assert.ifError(err);
-            renderer = _renderer;
-            renderer.getMetadata(this);
-        },
-        function(err, meta) {
-            if ( renderer ) {
-                renderer.release();
-            }
-            callback(err, meta);
-        }
-    );
-}
-
-function getMapnikLayerMetadata(layer, mapConfig, params, callback) {
-    step(
-        function getPGClient(err) {
-            assert.ifError(err);
-
-            var dbParams = RendererParams.dbParamsFromReqParams(params);
-            _.extend(dbParams, mapConfig.getLayerDatasource(params.layer));
-
-            return new PSQL(dbParams);
-        },
-        function fetchLayerTablenames(err, pg) {
-            assert.ifError(err);
-
-            var next = this;
-            var query = queryUtils.extractTableNames(layer.options.sql);
-
-            pg.query(query, function (err, result) {
-                if (err) {
-                    return next(err);
-                }
-
-                next(null, pg, result.rows[0].tablenames);
-            });
-        },
-
-        function getTableStats(err, pg, tablenames) {
-            assert.ifError(err);
-
-            var tableQueue = queue(tablenames.length);
-
-            tablenames.forEach(function (table) {
-                tableQueue.defer(function (pg, table, done) {
-                    var tableStats = {
-                        type: 'table',
-                        name: table,
-                    };
-                    var query = queryUtils.getTableStats(table);
-
-                    pg.query(query, function (err, resultTable) {
-                        if (err) {
-                            tableStats.features = -1;
-                        } else {
-                            try {
-                                tableStats.features = JSON.parse(resultTable.rows[0].result.stats).table_features;
-                            } catch (e) {
-                                tableStats.features = -1;
-                            }
-                        }
-                        done(null, tableStats);
-                    });
-                }, pg, table);
-            });
-
-            tableQueue.awaitAll(this);
-        },
-        function finnish(err, result) {
-            if (err) {
-                return callback(err);
-            }
-
-            callback(null, {
-                stats: result
-            });
-        }
-    );
-}
-
-function getEmptyLayerMetadada(callback) {
-    process.nextTick(function() {
-        callback(null, {
-            stats: []
-        });
-    });
-}
